dev: refactored layoutSlice
