+++ src/components/DappletSettings/DappletSettings.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import TagsGroup from 'components/TagsGroup'
 import { useResize } from 'hooks/useResize'
 import { useAppDispatch, useAppSelector } from 'store/hooks'
-import { setDappletSettingsState } from 'store/layoutSlice/layoutSlice'
+import { setDappletSettingsState } from 'store/slices/layoutSlice'
 import { addMyTag } from 'store/slices/myDappletsSlice'
 import CreateInput from 'uikit/CreateInput'
 import { SmartTagMode } from 'uikit/SmartTag/SmartTag'
