+++ src/components/Menu/Menu.tsx
@@ -4,7 +4,7 @@ import TagsGroup from 'components/TagsGroup'
 import { useResize } from 'hooks/useResize'
 import { nanoid } from 'nanoid'
 import { useAppDispatch, useAppSelector } from 'store/hooks'
-import { setMenuState } from 'store/layoutSlice/layoutSlice'
+import { setMenuState } from 'store/slices/layoutSlice'
 import MenuButton from 'uikit/MenuButton'
 import { MenuButtonIcon, MenuButtonMode } from 'uikit/MenuButton/MenuButton'
 import MyLists from 'uikit/MyLists'
